Split ServiceAlertsSection and TripViewControllerPreview into their own files
StopPageView.swift carried the whole service-alerts card — 144 lines of
a self-contained view that the page only references once — and
StopPageViewController.swift ended with TripViewControllerPreview, a
SwiftUI wrapper that has nothing to do with the view controller it was
sharing a file with. Both move out verbatim; no behavior changes.

Also adds a #Preview to StopPageHeaderView covering the three states the
pushed header can be in — with a walk time, without one, and the loading
placeholder — since the header is otherwise only reachable by running
the app and tapping through to a stop.
